update navigation bar appearance
